+++ src/main/java/io/confluent/idesidecar/restapi/kafkarest/controllers/ClusterManagerImpl.java
@@ -7,43 +7,37 @@ import static io.confluent.idesidecar.restapi.kafkarest.controllers.Relationship
 import static io.confluent.idesidecar.restapi.kafkarest.controllers.RelationshipUtil.forClusters;
 import static io.confluent.idesidecar.restapi.kafkarest.controllers.RelationshipUtil.forConsumerGroups;
 import static io.confluent.idesidecar.restapi.kafkarest.controllers.RelationshipUtil.forController;
-import static io.confluent.idesidecar.restapi.kafkarest.controllers.RelationshipUtil.forPartitionReassignments;
-import static io.confluent.idesidecar.restapi.kafkarest.controllers.RelationshipUtil.forPartitions;
-import static io.confluent.idesidecar.restapi.kafkarest.controllers.RelationshipUtil.forTopic;
-import static io.confluent.idesidecar.restapi.kafkarest.controllers.RelationshipUtil.forTopicConfigs;
 import static io.confluent.idesidecar.restapi.kafkarest.controllers.RelationshipUtil.forTopics;
 import static io.confluent.idesidecar.restapi.kafkarest.controllers.RelationshipUtil.getClusterCrn;
-import static io.confluent.idesidecar.restapi.kafkarest.controllers.RelationshipUtil.getTopicCrn;
 import static io.confluent.idesidecar.restapi.util.MutinyUtil.uniItem;
 import static io.confluent.idesidecar.restapi.util.MutinyUtil.uniStage;
+import static io.confluent.idesidecar.restapi.util.RequestHeadersConstants.CONNECTION_ID_HEADER;
 
+import io.confluent.idesidecar.restapi.cache.AdminClients;
 import io.confluent.idesidecar.restapi.exceptions.ClusterNotFoundException;
 import io.confluent.idesidecar.restapi.kafkarest.model.ClusterData;
 import io.confluent.idesidecar.restapi.kafkarest.model.ClusterDataList;
-import io.confluent.idesidecar.restapi.kafkarest.model.CreateTopicRequestData;
 import io.confluent.idesidecar.restapi.kafkarest.model.ResourceCollectionMetadata;
 import io.confluent.idesidecar.restapi.kafkarest.model.ResourceMetadata;
-import io.confluent.idesidecar.restapi.kafkarest.model.TopicData;
-import io.confluent.idesidecar.restapi.kafkarest.model.TopicDataList;
 import io.smallrye.mutiny.Uni;
+import io.vertx.core.http.HttpServerRequest;
+import jakarta.enterprise.context.RequestScoped;
+import jakarta.inject.Inject;
 import java.util.Collection;
 import java.util.List;
-import java.util.Optional;
-import java.util.Properties;
-import java.util.Set;
-import org.apache.kafka.clients.admin.AdminClient;
-import org.apache.kafka.clients.admin.DescribeTopicsOptions;
-import org.apache.kafka.clients.admin.NewTopic;
-import org.apache.kafka.clients.admin.TopicDescription;
+import java.util.function.Supplier;
 import org.apache.kafka.common.Node;
 
+@RequestScoped
 public class ClusterManagerImpl {
 
-  private final AdminClient adminClient;
+  @Inject
+  AdminClients adminClients;
 
-  public ClusterManagerImpl(Properties adminClientConfig) {
-    this.adminClient = AdminClient.create(adminClientConfig);
-  }
+  @Inject
+  HttpServerRequest request;
+
+  Supplier<String> connectionId = () -> request.getHeader(CONNECTION_ID_HEADER);
 
   public Uni<ClusterData> getKafkaCluster(String clusterId) {
     return describeCluster()
@@ -76,15 +70,15 @@ public class ClusterManagerImpl {
   }
 
   private Uni<ClusterDescribe> describeCluster() {
-      var describeCluster = adminClient.describeCluster();
-      return uniItem(new ClusterDescribe())
-          .chain(
-              cid -> uniStage(describeCluster.clusterId().toCompletionStage()).map(cid::withId))
-          .chain(cid -> uniStage(describeCluster.controller().toCompletionStage()).map(Node::id)
-              .map(cid::withControllerId)
-          ).chain(cid -> uniStage(describeCluster.nodes().toCompletionStage())
-              .map(cid::withNodes)
-          );
+    var describeCluster = adminClients.getAdminClient(connectionId.get()).describeCluster();
+    return uniItem(new ClusterDescribe())
+        .chain(
+            cid -> uniStage(describeCluster.clusterId().toCompletionStage()).map(cid::withId))
+        .chain(cid -> uniStage(describeCluster.controller().toCompletionStage()).map(Node::id)
+            .map(cid::withControllerId)
+        ).chain(cid -> uniStage(describeCluster.nodes().toCompletionStage())
+            .map(cid::withNodes)
+        );
   }
 
   private ClusterData fromClusterId(ClusterDescribe cluster) {
@@ -108,104 +102,6 @@ public class ClusterManagerImpl {
         .build();
   }
 
-  public Uni<TopicData> createKafkaTopic(String clusterId,
-      CreateTopicRequestData createTopicRequestData) {
-    return getKafkaCluster(clusterId)
-        .chain(ignored -> uniStage(
-            adminClient.createTopics(List.of(new NewTopic(
-                createTopicRequestData.getTopicName(),
-                Optional.ofNullable(createTopicRequestData.getPartitionsCount())
-                    .orElse(1),
-                Optional.ofNullable(createTopicRequestData.getReplicationFactor())
-                    .orElse(1).shortValue()
-            ))).all().toCompletionStage()))
-        .chain(v -> getKafkaTopic(
-            clusterId,
-            createTopicRequestData.getTopicName(),
-            false
-        ));
-  }
-
-  public Uni<Void> deleteKafkaTopic(String clusterId, String topicName) {
-    return getKafkaCluster(clusterId).chain(ignored ->
-        uniStage(
-            adminClient
-                .deleteTopics(List.of(topicName))
-                .all()
-                .toCompletionStage())
-    );
-  }
-
-  public Uni<TopicData> getKafkaTopic(
-      String clusterId, String topicName, Boolean includeAuthorizedOperations
-  ) {
-    var describeTopicsOptions = new DescribeTopicsOptions()
-        .includeAuthorizedOperations(
-            Optional.ofNullable(includeAuthorizedOperations).orElse(false)
-        );
-    return getKafkaCluster(clusterId).chain(ignored -> uniStage(
-        adminClient
-            .describeTopics(List.of(topicName), describeTopicsOptions)
-            .allTopicNames()
-            .toCompletionStage()
-    )
-        .map(topicDescriptions -> topicDescriptions.values().iterator().next())
-        .onItem().transform(topicDescription -> fromTopicDescription(clusterId, topicDescription)));
-  }
-
-  public Uni<TopicDataList> listKafkaTopics(String clusterId) {
-    return getKafkaCluster(clusterId).chain(ignored -> uniStage(
-        adminClient.listTopics().names().toCompletionStage()
-    ).chain(topicNames -> uniStage(
-            adminClient
-                .describeTopics(topicNames)
-                .allTopicNames()
-                .toCompletionStage())
-        ).onItem()
-        .transformToUni(topicDescriptionMap -> uniItem(TopicDataList
-            .builder()
-            .kind("KafkaTopicList")
-            .metadata(ResourceCollectionMetadata
-                .builder()
-                .next(null)
-                .self(forTopics(clusterId).getRelated())
-                .build()
-            )
-            .data(topicDescriptionMap
-                .values()
-                .stream()
-                .map(t -> fromTopicDescription(clusterId, t))
-                .toList()
-            ).build())
-        ));
-  }
-
-  private TopicData fromTopicDescription(String clusterId, TopicDescription topicDescription) {
-    return TopicData
-        .builder()
-        .kind("KafkaTopic")
-        .topicName(topicDescription.name())
-        .clusterId(clusterId)
-        .partitionsCount(topicDescription.partitions().size())
-        .replicationFactor(topicDescription.partitions().getFirst().replicas().size())
-        .isInternal(topicDescription.isInternal())
-        .authorizedOperations(
-            Optional.ofNullable(topicDescription.authorizedOperations()).orElse(Set.of())
-                .stream().map(Enum::name).toList()
-        )
-        .partitionReassignments(forPartitionReassignments(clusterId, topicDescription.name()))
-        .partitions(forPartitions(clusterId, topicDescription.name()))
-        .configs(forTopicConfigs(clusterId, topicDescription.name()))
-        .metadata(getTopicMetadata(clusterId, topicDescription.name())).build();
-  }
-
-  private ResourceMetadata getTopicMetadata(String clusterId, String topicName) {
-    return ResourceMetadata
-        .builder()
-        .resourceName(getTopicCrn(clusterId, topicName).toString())
-        .self(forTopic(clusterId, topicName).getRelated())
-        .build();
-  }
 
   private record ClusterDescribe(String id, Integer controllerId, Collection<Node> nodes) {
     ClusterDescribe() {
